shop/requirements.txt
amqp==5.2.0
asgiref==3.7.2
async-timeout==4.0.3
billiard==4.2.1
Brotli==1.1.0
cairocffi==1.7.1
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cryptography==43.0.3
cssselect2==0.7.0
defusedxml==0.8.0rc2
Django==5.0.9
django-debug-toolbar==4.4.6
django-extensions==3.2.3
django-localflavor==4.0
django-parler==2.3
django-rosetta==0.10.0
django-taggit==6.1.0
easy-thumbnails==2.10
flower==2.0.1
fonttools==4.54.1
html5lib==1.1
humanize==4.11.0
idna==3.10
kombu==5.4.2
Markdown==3.7
more-itertools==10.5.0
oauthlib==3.2.2
pillow==10.3.0
polib==1.2.0
prometheus_client==0.21.0
prompt_toolkit==3.0.48
psycopg2-binary==2.9.10
pycparser==2.22
pydyf==0.10.0
PyJWT==2.9.0
pyphen==0.16.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-stdnum==1.20
python3-openid==3.2.0
pytz==2024.2
redis==5.0.4
requests==2.32.3
requests-oauthlib==2.0.0
six==1.16.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.0
stripe==9.3.0
tinycss2==1.3.0
tornado==6.4.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
weasyprint==61.2
webencodings==0.5.1
zopfli==0.2.3
